backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# File handling
python-multipart==0.0.6
aiofiles==23.2.1

# MinIO/S3 integration
minio==7.2.0

# Configuration and environment
python-dotenv==1.0.0

# HTTP requests (optional, for testing)
requests==2.31.0

# Logging and monitoring
loguru==0.7.2

# Data validation
email-validator==2.1.0

# Security (optional)
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Development dependencies (install with pip install -r requirements-dev.txt)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# isort==5.12.0
# flake8==6.1.0 